Refactor helpers for seq server
