+++ packages/agents/sequencer/src/bindings/subscriber/index.ts
@@ -1,2 +1,1 @@
 export { bindSubscriber } from "./auctions";
-export { bindHealthServer } from "./server";
