+++ packages/agents/sequencer/src/index.ts
@@ -1,6 +1,6 @@
-import { makePublisher, makeSubscriber, makeSequencer, makeHTTPSubscriber } from "./make";
+import { makeHTTPPublisher, makeSubscriber, makeSequencer, makeHTTPSubscriber } from "./make";
 
 export { SequencerConfig, ChainConfig } from "./lib/entities";
-export { makePublisher, makeSubscriber, makeSequencer, makeHTTPSubscriber };
+export { makeHTTPPublisher, makeSubscriber, makeSequencer, makeHTTPSubscriber };
 
 makeSequencer();
